Issue #35 : Ponistavanje kupljene ulaznice - Metoda browseEvents() koja se nalazila u klasi User.java premjestena je u klasu RegularUser.java - U klasi RegularUser dodat je atribut ArrayList<Ticket> purchased_tickets koja sadrzi listu kupljenih ulaznica za datog korisnika - U klasi RegularUser dodata je metoda cancelPurchasedTicket() koja omogucava korisniku da ponisti kupljenu ulaznicu
